moving js files into corresponding folders
