+++ components/logging.js
